% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus.R
\name{minus}
\alias{minus}
\title{Return the difference of two numbers}
\usage{
minus(a, b)
}
\arguments{
\item{a}{Number one}

\item{b}{Number two}
}
\value{
the difference of two numbers
}
\description{
Return the difference of two numbers
}
\examples{
library(XmcPackage)
minus(35,12)
}
