% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{x
Just test Number(Description)}
\format{
An object of class \code{integer} of length 1000.
}
\usage{
x
}
\description{
x
Just test Number(Description)
}
\examples{
head(x)

}
\keyword{datasets}
